Refactor MockServer
After a second look of this code I realized that a refactor could be
made. Instead of using the functions: setStatusCode or setHeaders, it
would be better to code directly the handler as it is usually done in
express.

Another important refactor of MockServer is that now, the port where it
listen is zero. When the port is zero, the OS assigns a arbitrary free
port, this is more suitable for automated tests. Also I added a few
tests cases for the MockServer itself.

Finally and not less important, instances of Mock Servers are created
and destroyed after every test case. This is done to be able to code
different behaviour for the same path. In other words when you do:

mockserver.get("/", (req, res) => { // your code here })

If you want to change the behaviour for the same path in the same
instance of the MockServer you can't do it, since express already
registered the first handler. In order to overcome this I create a new
instance in every test execution